inspect_eval/pyproject.toml
[project]
name = "sysrepair-bench-inspect"
version = "0.1.0"
description = "Inspect AI eval harness for SysRepair-Bench scenarios"
requires-python = ">=3.11"
dependencies = [
    "inspect-ai>=0.3.70",
    "openai>=1.50.0",
    "pydantic>=2.7",
    "anyio>=4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tabulate>=0.9",
    "matplotlib>=3.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sysrepair_bench"]
